requirements.txt
discord-py-interactions==4.1.1b1
requests==2.26.0
python-dotenv==0.19.1
aiohttp[speedups]
